third_party/WebKit/LayoutTests/http/tests/inspector-enabled/open-with-emulation-enabled-expected.txt

This test checks that loading with emulation turned on by default loads elements module.

Loaded modules:
    main
    elements
    console
